+++ examples/humanoid_one_step.py
@@ -22,7 +22,7 @@ from dataclasses import dataclass
 import numpy as np
 import pylab
 
-from ltv_mpc import MPCProblem, solve_mpc
+from qpmpc import MPCProblem, solve_mpc
 
 
 @dataclass